% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counterButton}
\alias{counterButton}
\title{Counter Shiny Module UI}
\usage{
counterButton(id, label = "Counter +1")
}
\arguments{
\item{id}{(\code{character}) to describe the module}

\item{label}{(\code{character}) label on top of the button}
}
\description{
Counter Shiny Module UI
}
