% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_data.R
\docType{data}
\name{experimental_data}
\alias{experimental_data}
\title{Experimental Data}
\format{
An object of class \code{"list"};
}
\usage{
data(experimental_data)
}
\description{
Random data generated to simulate an antibody measurement of IgG and IgA antibodies.
}
\examples{
data(experimental_data)
print(experimental_data$IgG)
}
\keyword{datasets}
