% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_plot}
\alias{summary_plot}
\title{Create a grouped distribution plot}
\usage{
summary_plot(x, y)
}
\arguments{
\item{x}{(\code{numeric}) Vector with numbers}

\item{y}{(\code{factor}) Vector with factor levels}
}
\description{
Create a grouped distribution plot
}
\examples{
set.seed(1)
x <- sample(1:100, 50, replace = TRUE)
y <- as.factor(sample(c("Drug", "Placebo", "no treament"), 50, replace = TRUE))

summary_plot(x, y)

set.seed(1)
x <- c(
  rnorm(50, 115, sd = 25),
  rnorm(50, 198, sd = 65),
  sample(rnorm(50, 195, sd = 45), 50, TRUE)
)

y <- as.factor(c(
 rep("Drug", 50),
 rep("Placebo", 50),
 rep("no treatment", 50)
))

summary_plot(x, y)
}
