% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModule.R
\name{summaryServer}
\alias{summaryServer}
\title{Counter Shiny Module server}
\usage{
summaryServer(input, output, session, dataset)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny}

\item{dataset}{(\code{list}) A list of data.frames with two columns}
}
\value{
a number
}
\description{
Counter Shiny Module server
}
