% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModule.R
\name{summaryUI}
\alias{summaryUI}
\title{Counter Shiny Module UI}
\usage{
summaryUI(id, dataset = list())
}
\arguments{
\item{id}{(\code{character}) to describe the module}

\item{dataset}{(\code{list}) A list of data.frames with two columns}
}
\description{
Counter Shiny Module UI
}
